feat(chart): added chart component
